+++ src/app/pages/extinguishers/extinguisher-list/extinguisher-list.component.ts
@@ -2,12 +2,13 @@ import { Component, OnInit } from '@angular/core';
 import { Extinguisher } from 'src/app/models/extinguisher';
 import { ExtinguishersService } from '../extinguisher.service';
 import { Router } from '@angular/router';
-import { MatSnackBar } from '@angular/material';
+import { MatSnackBar, MatDialog } from '@angular/material';
 import { BaseListComponent } from 'src/app/commons/base-list-component';
 import { CustomersService } from '../../customers/customers.service';
 import { Condition } from 'src/app/commons/directives/filterable.directive';
 import { CustomSnackService } from 'src/app/services/custom-snack.service';
 import { environment } from 'src/environments/environment';
+import { DeletePopupComponent } from 'src/app/commons/delete-popup/delete-popup.component';
 
 @Component({
   selector: 'app-extinguisher-list',
@@ -16,18 +17,21 @@ import { environment } from 'src/environments/environment';
 })
 export class ExtinguisherListComponent extends BaseListComponent<Extinguisher> implements OnInit {
 
-  displayedColumns: string[] = ['code', 'customer', 'type', 'category', 'location', 'costCenter', 'address', 'factoryNo', 'bvNo', 'manufacturingDate', 'lastLoad', 'lastHydraulicTest', 'edit'];
+  displayedColumns: string[] = ['code', 'customer', 'type', 'category', /* 'location', 'costCenter', 'address', */ 'factoryNo', 'bvNo', 'manufacturingDate', 'lastLoad', 'lastHydraulicTest', 'edit'];
   dataSource: Extinguisher[];
 
   //Filters
   codeFilter: string = null;
   customerFilter: number;
+  factoryFilter: string = null;
+  clean: boolean = false;
 
   constructor(
     public service: ExtinguishersService, 
     protected router: Router,
     private _snackBar: CustomSnackService,
-    private customerService: CustomersService) { 
+    private customerService: CustomersService,
+    private dialog: MatDialog) { 
       super(service)
     }
 
@@ -53,16 +57,25 @@ export class ExtinguisherListComponent extends BaseListComponent<Extinguisher> i
   }
 
   delete(extType: any) : void {
-    this.service.delete(extType.id).subscribe(
-      _ => {
-        this._snackBar.showSuccess("Matafuego borrado!");
-        // this.loadData();
-        this.service.search()
-      },
-      error => {
-        this._snackBar.showError("Error borrando matafuego!");
+    const dialogRef = this.dialog.open(DeletePopupComponent, {
+      width: '400px'
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        this.service.delete(extType.id).subscribe(
+          _ => {
+            this._snackBar.showSuccess("Matafuego borrado!");
+            // this.loadData();
+            this.service.search()
+          },
+          error => {
+            this._snackBar.showError("Error borrando matafuego!");
+          }
+        )
       }
-    )
+    })
+    
   }
 
   create() : void {
@@ -79,15 +92,24 @@ export class ExtinguisherListComponent extends BaseListComponent<Extinguisher> i
     if(this.codeFilter && this.codeFilter != "")
       filters.push({column: 'code', condition: Condition.contains, value: this.codeFilter}) 
     if(this.customerFilter)
-      filters.push({column: 'customer', condition: Condition["="], value: this.customerFilter}) 
+      filters.push({column: 'customer', condition: Condition["="], value: this.customerFilter})
+    if(this.factoryFilter && this.factoryFilter != "")
+      filters.push({column: 'factoryNo', condition: Condition["="], value: this.factoryFilter}) 
     if(filters.length > 0)
       this.service.filters = filters;
     this.service.search()
   }
 
   selectCustomer(id: any) : void {
-    console.log("Selected:", id)
+    this.clean = false;
     this.customerFilter = (!id || id == 'null') ? null : id;
+  }
+
+  cleanFilter() : void {
+    this.codeFilter = null;
+    this.factoryFilter = null;
+    this.customerFilter = null;
+    this.clean = true;
   }
 
 }
